backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Vector database for RAG
chromadb==0.4.22

# LLM integration
ollama==0.1.6
httpx==0.26.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3
html2text==2024.2.26

# File watching
watchdog==3.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
uuid7==0.1.0

# Async support
aiofiles==23.2.1

# Text processing
tiktoken==0.5.2
langdetect==1.0.9
